move test utils to their own file
